requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
decorator==5.1.1
distro==1.9.0
executing==0.8.3
fastapi==0.115.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
numpy==2.2.3
openai==2.9.0
parso==0.8.4
pexpect==4.8.0
pip==25.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pvrhino==4.0.0
PyAudio==0.2.14
pycparser==2.23
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
RPi.GPIO==0.7.1
setuptools==75.8.0
sniffio==1.3.1
sounddevice==0.5.3
srt==3.5.3
stack-data==0.2.0
starlette==0.46.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.34.0
vosk==0.3.45
wcwidth==0.2.5
websockets==15.0.1
webrtcvad==2.0.10
wheel==0.45.1
